+++ src/logger/loggers/sql.ts
@@ -1,10 +1,11 @@
-import { transports, createLogger, format } from 'winston';
-import 'winston-daily-rotate-file';
-
 import {
   LOG_FILENAME_SQL, LOG_DATE_PATTERNT, LOG_MAX_SIZE, LOG_MAX_FILES,
 } from '../../utils/constants';
+import { Winston } from '../utils/configureLogger';
+
 import loggerFormatter from '../utils/logFormatter';
+
+const { createLogger, transports, format } = Winston;
 
 export default (config: Config) => {
   const { logDir } = config;