+++ src/types.ts
@@ -3,7 +3,7 @@ import http from 'http';
 import https from 'https';
 import DeviceDetector from 'device-detector-js';
 import { NextFunction, Request, Response } from 'express';
-import { GraphQLSchema } from 'graphql';
+import { GraphQLSchema, DocumentNode } from 'graphql';
 import { IMiddleware } from 'graphql-middleware';
 import { RedisPubSub } from 'graphql-redis-subscriptions';
 import { ITypedef, IResolvers } from 'graphql-tools';
@@ -11,12 +11,16 @@ import { RedisOptions, Redis as RedisInterface } from 'ioredis';
 import { Options as SesstionStoreOptions } from 'session-file-store';
 import { ServerOptions as IWebsocketServerOption } from 'ws';
 
+import './graphql-ext';
 import { IDBConfig, Knex } from './databaseManager';
 import { ILoggerCollection } from './logger';
 import {
   IJwtConfig,
   IAccessToken,
 } from './schemas/auth/types';
+
+export * from 'graphql-tools';
+
 
 export interface IInitProps {
   port?: number;
@@ -113,3 +117,5 @@ export type IExpressMidlewareContainer = (props: {context: IContext}) => IExpres
 
 export type IExpressMiddleware = (
   request?: Request, response?: Response, next?: NextFunction) => void;
+
+export type GraphQLExtension = DocumentNode;